+++ libs/proc/src/lib.rs
@@ -14,7 +14,6 @@ pub struct ProcId(pub usize);
 
 impl ProcId {
     pub const NULL: Self = Self(0);
-    pub const KERNEL: Self = Self(0);
 }
 
 #[derive(Debug, Hash, Eq, PartialEq, Ord, PartialOrd, Clone, Copy)]
@@ -22,48 +21,56 @@ pub struct TaskId(pub usize);
 
 impl TaskId {
     pub const NULL: Self = Self(0);
-    pub const KERNEL: Self = Self(0);
 }
 
 /// Process manager
 pub trait ProcessManager {
-    fn new_state(&self) -> Box<dyn Any>;
+    /// Create a new process
     fn spawn(&self, t: Box<dyn Runnable>, mm: Box<dyn Any>) -> Arc<dyn Proc>;
+    /// Get a process by its id
     fn get_proc_by_id(&self, id: ProcId) -> Option<Arc<dyn Proc>>;
+    /// Get the current process
     fn current_proc(&self) -> Option<Arc<dyn Proc>>;
+    /// Get the current process id
     fn current_proc_id(&self) -> Option<ProcId>;
+    /// Get a task by its id
     fn get_task_by_id(&self, id: TaskId) -> Option<Arc<dyn Task>>;
+    /// Get the current task
     fn current_task(&self) -> Option<Arc<dyn Task>>;
-}
-
-#[derive(Debug, Hash, Eq, PartialEq, Ord, PartialOrd, Clone, Copy)]
-pub enum TaskState {
-    Ready,
-    Running,
-    Blocked,
-    Dead,
 }
 
 /// Abstruct task type
 pub trait Task: Send + Sync {
+    /// Get task id
     fn id(&self) -> TaskId;
+    /// Get task running state
     fn state(&self) -> &Monitor<bool>;
+    /// Arch-dependent context for context switching
     fn context(&self) -> &dyn Any;
+    /// The process that owns this task
     fn proc(&self) -> Arc<dyn Proc>;
+    /// Task scheduler state
     fn sched(&self) -> &dyn Any;
+    /// The task runnable
     fn runnable(&self) -> &dyn Runnable;
 }
 
 /// Abstruct process type
 pub trait Proc: Send + Sync + Any {
+    /// Get process id
     fn id(&self) -> ProcId;
+    /// Get process vfs state
     fn fs(&self) -> &dyn Any;
+    /// Get process memory state
     fn mm(&self) -> &dyn Any;
+    /// Get all the tasks in this process
     fn tasks(&self) -> &Mutex<Vec<TaskId>>;
+    /// Spawn a task
     fn spawn_task(self: Arc<Self>, task: Box<dyn Runnable>) -> Arc<dyn Task>;
+    /// Exit the process
     fn exit(&self);
+    /// Wait for the process to complete
     fn wait_for_completion(&self);
-    // fn sbrk(&self, f: *const extern "C" fn());
 }
 
 /// Holds the execution code for a kernel task.
