--- modules/pm/src/user.rs
@@ -1,1 +1,0 @@
-
